requirements.txt
langchain>=0.1.0
ipykernel
python-dotenv
langchain-community 
pypdf
bs4
arxiv
pymupdf
wikipedia

langchain_text_splitters

sentence-transformers
langchain_huggingface

python-dotenv
langchain-openai
langchain-ollama
langchain-chroma>=0.1.2
chromadb>=0.3.23
faiss-cpu

beautifulsoup4
streamlit>=1.20.0
streamlit-chat
